# docs/swagger.yaml
basePath: /
definitions:
...
  organizations.CreateOrganizationRequest:
    properties:
      address_line1:
        example: 1 Main St
        maxLength: 255
        type: string
      address_line2:
        example: Suite 200
        maxLength: 255
        type: string
      billing_email:
        example: billing@acme.com
        maxLength: 320
        type: string
      city:
        example: Austin
        maxLength: 120
        type: string
      company_registration:
        example: DE-12345
        maxLength: 128
        type: string
      country:
        example: US
        type: string
      default_currency:
        example: USD
        type: string
      legal_name:
        example: Acme Property Management LLC
        maxLength: 255
        type: string
      name:
        example: Acme Property Management
        maxLength: 255
        minLength: 2
        type: string
      phone:
        example: +1-555-0100
        maxLength: 64
        type: string
      postal_code:
        example: "78701"
        maxLength: 32
        type: string
      region:
        example: TX
        maxLength: 120
        type: string
      slug:
        example: acme-prop
        maxLength: 128
        minLength: 2
        type: string
      tax_id:
        example: 12-3456789
        maxLength: 64
        type: string
      timezone:
        example: America/Chicago
        maxLength: 64
        type: string
      website:
        example: https://acme.com
        maxLength: 512
        type: string
    required:
    - name
    - slug
    type: object
  organizations.OrganizationResponse:
    properties:
      address_line1:
        type: string
      address_line2:
        type: string
      billing_email:
        type: string
      city:
        type: string
      company_registration:
        type: string
      country:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      default_currency:
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      legal_name:
        type: string
      name:
        example: Acme Property Management
        type: string
      phone:
        type: string
      postal_code:
        type: string
      region:
        type: string
      slug:
        example: acme-prop
        type: string
      tax_id:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      website:
        type: string
    type: object
  organizations.PatchOrganizationRequest:
    properties:
      address_line1:
        maxLength: 255
        type: string
      address_line2:
        maxLength: 255
        type: string
      billing_email:
        maxLength: 320
        type: string
      city:
        maxLength: 120
        type: string
      company_registration:
        maxLength: 128
        type: string
      country:
        type: string
      default_currency:
        type: string
      legal_name:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      phone:
        maxLength: 64
        type: string
      postal_code:
        maxLength: 32
        type: string
      region:
        maxLength: 120
        type: string
      tax_id:
        maxLength: 64
        type: string
      timezone:
        maxLength: 64
        type: string
      website:
        maxLength: 512
        type: string
    type: object
...
  response.Envelope-organizations_OrganizationResponse:
    properties:
      data:
        $ref: '#/definitions/organizations.OrganizationResponse'
    type: object
...
  response.ErrorBody:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
    type: object
  response.ErrorDetail:
    properties:
      code:
        type: string
      kind:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Rent Management System HTTP API (bootstrap; module routes are placeholders).
  title: RMS API
  version: "1.0"
paths:
...
  /api/v1/organizations:
    post:
      consumes:
      - application/json
      description: '**Admin only.** Creates a new landlord / property-management company
        (tenant). Sets `created_by` / `updated_by` from the access token subject.'
      parameters:
      - description: Organization payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/organizations.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Envelope-organizations_OrganizationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Non-admin callers are rejected
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create organization
      tags:
      - organizations
  /api/v1/organizations/{id}:
    get:
      description: '**Admin:** any organization id. **Landlord, manager, staff:**
        only their own `organization_id` from the token; cross-tenant access returns
        403.'
      parameters:
      - description: Organization ID
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope-organizations_OrganizationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get organization
      tags:
      - organizations
    patch:
      consumes:
      - application/json
      description: '**Admin:** may update any organization. **Landlord and manager:**
        may update only their own organization. **Staff** cannot use this endpoint.
        Slug is immutable. Sets `updated_by` from the token subject.'
      parameters:
      - description: Organization ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update (omit to leave unchanged)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/organizations.PatchOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope-organizations_OrganizationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update organization profile
      tags:
      - organizations
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"